wip: linking between notes and note entrypoints.
The note editor should now only be responsible for showing the entry point notes, all other notes should be accessible from links in notes or via the search.

Instead of using the cache id we now have a special endpoint for retrieving notes via the documentId which is the id stored in the database.

A note created from the note editor is automatically a entrypoint note. In a follow up commit it will be possible to create/link new notes inline within the markdown editor that will not appear as a entrypoint.

The notes created by linking a reference are treated as none entry-points.

A follow-up commit should allow configuring whether a note is a entrypoint or not via the UI in the application.